Privilege updates Add, Edit, Delete, View for each sector
git-svn-id: http://svn.openmrs.org@438 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ src/api/org/openmrs/api/AdministrationService.java
@@ -309,6 +309,8 @@ public class AdministrationService {
 	 * @throws APIException
 	 */
 	public void deleteRole(Role role) throws APIException {
+		if (OpenmrsConstants.CORE_ROLES().contains(role.getRole()))
+			throw new APIException("Cannot delete a core role");
 		getAdminDAO().deleteRole(role);
 	}
 
@@ -338,7 +340,7 @@ public class AdministrationService {
 	public void deletePrivilege(Privilege privilege) throws APIException {
 		if (OpenmrsConstants.CORE_PRIVILEGES().contains(privilege.getPrivilege()))
 			throw new APIException("Cannot delete a core privilege");
-		getAdminDAO().deletePrivilege(privilege);
+			getAdminDAO().deletePrivilege(privilege);
 	}
 
 	/**
